Lots of stuff has been added and fixed. Updated README.md

+++ src/BusinessLogicLayer/Services/StudentsService.cs
@@ -59,7 +59,24 @@ namespace BusinessLogicLayer.Services
             }
 
             ApplyMd5Hashing(ref Credential);
-            return _appContext.getUoW().StudentsRepository.LoginStudent(Credential);
+
+            StudentModel ReturnStd = _appContext.getUoW().StudentsRepository.LoginStudent(Credential);
+            SetFineAmounts(_appContext, ref ReturnStd);
+
+            return ReturnStd;
+        }
+
+        public static void SetFineAmounts(CustomAppContext AppContext, ref StudentModel Model) {
+            try
+            {
+                Model.total_fine = AppContext.getUoW().BorrowingRepository.GetTotalFineForStudent(Model.student_number);
+                Model.active_fine = AppContext.getUoW().BorrowingRepository.GetActiveFineForStudent(Model.student_number);
+            }
+            catch
+            {
+                Model.total_fine = 0;
+                Model.active_fine = 0;
+            }
         }
 
         public void AddStudent(StudentModel Student)
@@ -110,17 +127,39 @@ namespace BusinessLogicLayer.Services
 
         public void DeleteStudentById(int Id)
         {
-            StudentModel TempStudent = new StudentModel();
-            TempStudent.student_number = Id;
-            var output = _appContext.getUoW().StudentsRepository.Delete(TempStudent);
-            _appContext.getUoW().Commit();
+            int activeBookCount = _appContext.getUoW().BorrowingRepository.GetActiveBorrowingCountForStudent(Id);
+            if (activeBookCount > 0)
+            {
+                throw new Exception("Can't delete. Student has '" + activeBookCount + "' active book(s).");
+            }
+            else {
+                StudentModel TempStudent = new StudentModel();
+                TempStudent.student_number = Id;
+                var output = _appContext.getUoW().StudentsRepository.Delete(TempStudent);
+                var output2 = _appContext.getUoW().BorrowingRepository.UpdateStudent(Id);
+                _appContext.getUoW().Commit();
+            }
         }
 
         public void UpdateStudent(StudentModel Student)
         {
+            StudentsValidator validator = new StudentsValidator();
+            ValidationResult results = validator.Validate(Student);
+            string allMessages = results.ToString("\n");
+
+            if (!results.IsValid)
+            {
+                throw new Exception(allMessages);
+            }
+
             ApplyMd5Hashing(ref Student);
             var output = _appContext.getUoW().StudentsRepository.Update(Student);
             _appContext.getUoW().Commit();
+
+            if (output != 1)
+            {
+                throw new Exception("Error");
+            }
         }
 
         public void BulkAdd(List<StudentModel> Students)
